Changed lexer to handle all types of alfanumerical indicators. Now indicators OC, OF and OV are handled properly.
