+++ rpgJavaInterpreter-core/src/main/antlr/RpgLexer.g4
@@ -316,8 +316,10 @@ SPLAT_GETIN: '*'[gG][eE][tT][iI][nN];
 SPLAT_HIVAL: '*'[hH][iI][vV][aA][lL];
 SPLAT_INIT: '*'[iI][nN][iI][tT];
 SPLAT_INDICATOR : 	( '*' [iI] [nN] [0-9] [0-9]
-					| '*' [iI] [nN] [lL] [rR]
+					| '*' [iI] [nN] [a-zA-Z] [a-zA-Z]
 					| '*' [iI] [nN] [rR] [tT]
+					| '*' [iI] [nN] [oO] [vV]
+					| '*' [iI] [nN] [oO] [vF]
 					| '*' [iI] [nN] '(' [0-9] [0-9] ')' );
 SPLAT_INZSR: '*'[iI][nN][zZ][sS][rR];
 SPLAT_IN: '*'[iI][nN];
